requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
fastapi
streamlit
langsmith
langserve
uvicorn
python-dotenv
requests
pyttsx3
gtts
eSpeak